+++ frontend/ESCOM-view/src/ModuloDocumental/components/Matrix/Config/list/trainingArea.js
@@ -3,6 +3,7 @@ import { Component } from 'react';
 import { connect } from 'react-redux';
 import { withRouter } from 'react-router-dom';
 import { reduxForm, Field } from 'redux-form';
+import { confirmAlert } from 'react-confirm-alert';
 import { required, minimum, fiveHundred } from '../../../utilitarian/validations.js';
 import { getListGeneralC, addGeneralC, addMessageAdd, addMessageDelete, deleteGeneralC } from '../../../../redux/actions/generalClassA.js';
 
@@ -56,14 +57,29 @@ class TrainingArea extends Component {
 
 
     disable(id) {
-        let generalA = {
-            id: id,
-            name: "",
-            idGeneral: "",
-            table: 'TrainingArea',
-            requestData: null
-        }
-        this.props.deleteGeneralC(localStorage.getItem('Token'), generalA)
+        confirmAlert({
+            title: '',
+            message: '¿Esta seguro?',
+            buttons: [
+                {
+                    label: 'Si',
+                    onClick: () => {
+                        let generalA = {
+                            id: id,
+                            name: "",
+                            idGeneral: "",
+                            table: 'TrainingArea',
+                            requestData: null
+                        }
+                        this.props.deleteGeneralC(localStorage.getItem('Token'), generalA)
+                    }
+                },
+                {
+                    label: 'No',
+                    onClick: () => { }
+                }
+            ]
+        });
     }
 
     render() {
